requirements.txt
bse==3.1.0
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
pandas==2.2.0
loguru==0.7.2
APScheduler==3.10.4
python-dotenv==1.0.1
pyyaml==6.0.1
psycopg2-binary==2.9.9
pydantic==2.6.0
pydantic-settings==2.1.0

# dbt dependencies - matching versions
dbt-core==1.5.0
dbt-sqlite==1.5.1
